src/main/java/Vending_Machine/Product.java
package Vending_Machine;

public class Product {
    private String id;
    private String name;
    private Category category;
    private double price;
    private int quantity;
    private int quantitySold;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Category getCategory() {
        return category;
    }

    public void setCategory(Category category) {
        this.category = category;
    }

    public double getPrice() {
        return price;
    }

    public void setPrice(double price) {
        this.price = price;
    }

    public int getQuantity() {
        return quantity;
    }

    public void setQuantity(int quantity) {
        this.quantity = quantity;
    }

    public int getQuantitySold() {
        return quantitySold;
    }

    public void setQuantitySold(int quantitySold) {
        this.quantitySold = this.quantitySold + quantitySold;
    }

    public Product(String id, String name, Category category, double price, int quantity, int quantitySold) {
        this.id = id;
        this.name = name;
        this.category = category;
        this.price = price;
        this.quantity = quantity;
        this.quantitySold = quantitySold;
    }
}
